Creates a label of arbitrary content
Source: R/Class-Label.R

See https://docs.docker.com/engine/reference/builder/#label

Label(..., multi_line = FALSE, label_ns = NA_character_)

Arguments

...
One or more named arguments (key-value pairs), that the Instruction should consist of

multi_line
Whether to print labels with multiple values in one single line or to put each key-value pair in a new line.

label_ns
Optionally specify a namespace that should be appended to all keys of the label

Value

the object

See also

Other label: Label-class, LabelSchemaFactory, Label_Maintainer, Label_SessionInfo

Examples

# A simple label that occupies one line:
label1 <- Label(key1 = "this", key2 = "that", otherKey = "content")
# A multiline label with one key/value pair per line
label2 <- Label(
 "info.o2r.name" = "myProject_ImageName",
 "org.label-schema.name" = "ImageName",
 "yet.another_labelname" = "true",
 multi_line = TRUE
)
#label with fixed namespace for all keys
label3 <- Label("name"="A name", "description" = "A description", label_ns = "my.label.ns.")
df <- dockerfile(clean_session())

#> INFO [2018-03-26 22:21:51] Creating an R session with the following arguments:
#> 	 R  --silent --vanilla -e "info <- sessionInfo()" -e "save(list = \"info\", file = \"/tmp/RtmpBJq2xz/rdata-sessioninfo19e52ee35ce3\")"
#> INFO [2018-03-26 22:21:53] Created Dockerfile-Object based on sessionInfo

addInstruction(df) <- list(label1, label2, label3)
cat(format(df),sep = "\n")

#> FROM rocker/r-ver:3.4.4                                                                                             
#> LABEL maintainer="daniel"                                                                                           
#> WORKDIR /payload/                                                                                                   
#> LABEL key1="this" key2="that" otherKey="content"                                                                    
#> LABEL info.o2r.name="myProject_ImageName" \
#> 	org.label-schema.name="ImageName" \
#> 	yet.another_labelname="true"
#> LABEL my.label.ns.name="A name" my.label.ns.description="A description"                                             
#> CMD ["R"]