dockerfile-method
Source: R/dockerfile.R

Create a Dockerfile based on either a sessionInfo, a workspace or a file.

dockerfile(from = utils::sessionInfo(),
  image = getImageForVersion(getRVersionTag(from)),
  maintainer = Sys.info()[["user"]], save_image = FALSE,
  env = list(generator = paste("containerit",
  utils::packageVersion("containerit"))), soft = FALSE, offline = FALSE,
  copy = NA, container_workdir = "/payload/", cmd = Cmd("R"),
  entrypoint = NULL, add_self = FALSE, vanilla = TRUE, silent = FALSE,
  predetect = TRUE, versioned_libs = FALSE, versioned_packages = FALSE,
  filter_baseimage_pkgs = FALSE)

Arguments

from
The source of the information to construct the Dockerfile. Can be a sessionInfo object, a path to a file, or the path to a workspace). If NULL then no automatic derivation of dependencies happens.

image
(From-object or character) Specifes the image that shall be used for the Docker container (FROM instruction). By default, the image is determinded from the given session. Alternatively, use getImageForVersion(..) to get an existing image for a manually defined version of R, matching the version with tags from the base image rocker/r-ver (see details about the rocker/r-ver at 'https://hub.docker.com/r/rocker/r-ver/'). Or provide a correct image name yourself.

maintainer
Specify the maintainer of the dockerfile. See documentation at 'https://docs.docker.com/engine/reference/builder/#maintainer'. Defaults to Sys.info()[["user"]]. Can be removed with NULL.

save_image
When TRUE, it calls save.image and include the resulting .RData in the container's working directory. Alternatively, you can pass a list of objects to be saved, which may also include arguments to be passed down to save. E.g. save_image = list("object1", "object2", file = "path/in/wd/filename.RData"). save will be called with default arguments file = ".RData" and envir = .GlobalEnv

env
optionally specify environment variables to be included in the image. See documentation: 'https://docs.docker.com/engine/reference/builder/#env

soft
(boolean) Whether to include soft dependencies when system dependencies are installed, default is no.

offline
(boolean) Whether to use an online database to detect system dependencies or use local package information (slower!), default is no.

copy
whether and how a workspace should be copied - values: "script", "script_dir" or a list of relative file paths to be copied, or NA ot disable copying of files

container_workdir
the working directory in the container, defaults to /payload/ and must end with /. Can be skipped with value NULL.

cmd
The CMD statement that should be executed by default when running a parameter. Use cmd_Rscript(path) in order to reference an R script to be executed on startup

entrypoint
the ENTRYPOINT statement for the Dockerfile

add_self
Whether to add the package containerit itself if loaded/attached to the session

vanilla
Whether to use an empty vanilla session when packaging scripts and markdown files (equivalent to R --vanilla)

silent
Whether or not to print information during execution

predetect
Extract the required libraries based on library calls using the package automagic before running a script/document

versioned_libs
[EXPERIMENTAL] Whether it shall be attempted to match versions of linked external libraries

versioned_packages
[EXPERIMENTAL] Whether it shall be attempted to match versions of R packages

filter_baseimage_pkgs
Whether it shall be attempted to match versions of R packages

Value

An object of class Dockerfile

Based on sessionInfo

Use the current sessionInfo) to create a Dockerfile.

Based on a workspace/directory

Given an existing path to a directory, the method tries to automatically find the main R file within that directory. Files are searched recursively. The following types are supported:

regular R script files, identified by file names ending in .R

weaved documents, identified by file names ending in R Markdown (.Rmd)

After identifying the main file, the process continues as described in the section file. If both types are found, documents are given priority over scripts. If multiple files are found, the first file as returned by dir will be used.

Based on a file

Given an executable R script or document, create a Dockerfile to execute this file. This executes the whole file to obtain a complete sessionInfo object, see section "Based on sessionInfo", and copies required files and documents into the container.

Examples

dockerfile <- dockerfile()

#> Warning: DESCRIPTION file of package 'plm' is missing or broken

#> Error: $ operator is invalid for atomic vectors

print(dockerfile)

#> function(from = utils::sessionInfo(),
#>                        image = getImageForVersion(getRVersionTag(from)),
#>                        maintainer = Sys.info()[["user"]],
#>                        save_image = FALSE,
#>                        env = list(generator = paste("containerit", utils::packageVersion("containerit"))),
#>                        soft = FALSE,
#>                        offline = FALSE,
#>                        copy = NA, # "script",
#>                        # nolint start
#>                        container_workdir = "/payload/",
#>                        # nolint end
#>                        cmd = Cmd("R"),
#>                        entrypoint = NULL,
#>                        add_self = FALSE,
#>                        vanilla = TRUE,
#>                        silent = FALSE,
#>                        predetect = TRUE,
#>                        versioned_libs = FALSE,
#>                        versioned_packages = FALSE,
#>                        filter_baseimage_pkgs = FALSE) {
#>     if (silent) {
#>       invisible(futile.logger::flog.threshold(futile.logger::WARN))
#>     }
#> 
#>     .dockerfile <- NA
#>     .originalFrom <- class(from)
#> 
#>     #parse From-object from string if necessary
#>     if (is.character(image)) {
#>       image <- parseFrom(image)
#>     }
#>     futile.logger::flog.debug("Creating a new Dockerfile from '%s' with base image %s", from, toString(image))
#> 
#>     if (is.character(maintainer)) {
#>       .label <- Label_Maintainer(maintainer)
#>       futile.logger::flog.debug("Turning maintainer character string '%s' into label: %s", maintainer, toString(.label))
#>       maintainer <- .label
#>     }
#> 
#>     # check CMD instruction
#>     if (!inherits(x = cmd, "Cmd")) {
#>       stop("Unsupported parameter for 'cmd', expected an object of class 'Cmd', given was :",
#>         class(cmd))
#>     }
#> 
#>     # check ENTRYPOINT instruction
#>     if ( !is.null(entrypoint) && !inherits(x = entrypoint, "Entrypoint")) {
#>       stop("Unsupported parameter for 'entrypoint', expected an object of class 'Entrypoint', given was :",
#>         class(entrypoint))
#>     }
#> 
#>     # check and create WORKDIR instruction
#>     workdir <- NULL
#>     if ( !is.null(container_workdir)) {
#>       if ( !is.character(container_workdir)) {
#>         stop("Unsupported parameter for 'container_workdir', expected a character string or NULL")
#>       } else {
#>         # nolint start
#>         if (!stringr::str_detect(container_workdir, "/$")) {
#>           # nolint end
#>           # directories given as destination must have a trailing slash in dockerfiles
#>           container_workdir <- paste0(container_workdir, "/")
#>           futile.logger::flog.info("Appended trailing slash, workdir is '%s'", container_workdir)
#>         }
#>         workdir <- Workdir(container_workdir)
#>       }
#>     }
#> 
#>     # whether image is supported
#>     image_name <- image@image
#>     if (!image_name %in% .supported_images) {
#>       warning("Unsupported base image. Proceed at your own risk. The following base images are supported:\n",
#>         paste(.supported_images, collapse = "\n"))
#>     }
#> 
#>     # base dockerfile
#>     .dockerfile <- methods::new("Dockerfile",
#>                                 instructions = list(),
#>                                 maintainer = maintainer,
#>                                 image = image,
#>                                 entrypoint = entrypoint,
#>                                 cmd = cmd)
#> 
#>     # handle differe "from" cases
#>     if (is.null(from)) {
#>       futile.logger::flog.debug("from is NULL, not deriving any information at all")
#>       if (!is.null(workdir))
#>         addInstruction(.dockerfile) <- workdir
#>     } else if (inherits(x = from, "sessionInfo")) {
#>       futile.logger::flog.debug("Creating from sessionInfo object")
#>       .dockerfile <- dockerfileFromSession(.dockerfile = .dockerfile,
#>                                            session = from,
#>                                            soft = soft,
#>                                            offline = offline,
#>                                            add_self = add_self,
#>                                            versioned_libs = versioned_libs,
#>                                            versioned_packages = versioned_packages,
#>                                            filter_baseimage_pkgs = filter_baseimage_pkgs,
#>                                            workdir = workdir)
#>     } else if (inherits(x = from, "character")) {
#>       futile.logger::flog.debug("Creating from character string '%s'", from)
#> 
#>       if (dir.exists(from)) {
#>         futile.logger::flog.debug("'%s' is a directory", from)
#>         .originalFrom <- from
#>         .dockerfile <- dockerfileFromWorkspace(path = from,
#>                                                .dockerfile = .dockerfile,
#>                                                soft = soft,
#>                                                offline = offline,
#>                                                add_self = add_self,
#>                                                copy = copy,
#>                                                vanilla = vanilla,
#>                                                silent = silent,
#>                                                predetect = predetect,
#>                                                versioned_libs = versioned_libs,
#>                                                versioned_packages = versioned_packages,
#>                                                filter_baseimage_pkgs = filter_baseimage_pkgs,
#>                                                workdir = workdir)
#>       } else if (file.exists(from)) {
#>         futile.logger::flog.debug("'%s' is a file", from)
#>         .originalFrom <- from
#>         .dockerfile <- dockerfileFromFile(file = from,
#>                                           .dockerfile = .dockerfile,
#>                                           soft = soft,
#>                                           offline = offline,
#>                                           add_self = add_self,
#>                                           copy = copy,
#>                                           vanilla = vanilla,
#>                                           silent = silent,
#>                                           predetect = predetect,
#>                                           versioned_libs = versioned_libs,
#>                                           versioned_packages = versioned_packages,
#>                                           filter_baseimage_pkgs = filter_baseimage_pkgs,
#>                                           workdir = workdir)
#>       } else {
#>         stop("Unsupported string for 'from' argument (not a file, not a directory): ", from)
#>       }
#>     } else if (is.expression(from) ||
#>                (is.list(from) && all(sapply(from, is.expression)))) {
#>       futile.logger::flog.debug("Creating from expession: '%s' with a clean session", toString(from))
#> 
#>       .sessionInfo <- clean_session(expr = from,
#>                                     slave = silent,
#>                                     vanilla = vanilla)
#>       .dockerfile <- dockerfileFromSession(.dockerfile = .dockerfile,
#>                                            session = .sessionInfo,
#>                                            soft = soft,
#>                                            offline = offline,
#>                                            add_self = add_self,
#>                                            versioned_libs = versioned_libs,
#>                                            versioned_packages = versioned_packages,
#>                                            filter_baseimage_pkgs = filter_baseimage_pkgs,
#>                                            workdir = workdir)
#>     } else {
#>       stop("Unsupported 'from': ", class(from), " ", from)
#>     }
#> 
#>     # copy any additional files / objects into the working directory from here:
#>     if (isTRUE(save_image)) {
#>       save.image()
#>       addInstruction(.dockerfile) <-
#>         Copy(src = "./.RData", dest = "./")
#>     } else if (is.list(save_image)) {
#>       do.call(.save_objects, save_image)
#>       if ("file" %in% names(save_image)) {
#>         file <- save_image$file
#>         # try to assure unix-compatibility..
#>         file <- stringr::str_replace_all(file, "\\\\", "/")
#>       } else
#>         file = "./payload.RData"
#>       addInstruction(.dockerfile) <- Copy(src = file, dest = file)
#>     }
#> 
#>     futile.logger::flog.info("Created Dockerfile-Object based on %s", .originalFrom)
#>     return(.dockerfile)
#>   }
#> <environment: namespace:containerit>