Write a Dockerfile object to a dockerfile
Source: R/Class-Dockerfile.R

Write a Dockerfile object to a dockerfile

# S4 method for Dockerfile
write(x, file = file.path(getwd(), "Dockerfile"))

Arguments

x
Dockerfile object to be serialized

file
optional argument specifying a costum file path

Examples

# write a dockerfile with default parameters to temporary file and show content:
temp = tempfile()
write(dockerfile(), file=temp)

#> Warning: Failed to identify a source for package containerit. Therefore the package cannot be installed in the Docker image.

#> Warning: Failed to identify a source for package harbor. Therefore the package cannot be installed in the Docker image.

#> INFO [2018-03-26 22:22:17] Going online? TRUE  ... to retrieve system dependencies (sysreq-api)
#> INFO [2018-03-26 22:22:17] Trying to determine system requirements for the package(s) 'A3,pbapply,xtable,abc,locfit,MASS,quantreg,SparseM,nnet,abc.data,sessioninfo,devtools,testthat,httr,pkgload,jsonlite,assertthat,yaml,highlight,pillar,backports,lattice,glue,digest,colorspace,htmltools,Matrix,plyr,clisymbols,pkgconfig,purrr,scales,whisker,MatrixModels,tibble,pacman,withr,fortunes,cli,magrittr,crayon,memoise,evaluate,fs,xml2,githubinstall,pkgbuild,rsconnect,data.table,formatR,stringr,sysreqs,munsell,lambda.r,bindrcpp,callr,pkgdown,rlang,futile.logger,debugme,rstudioapi,sys,rjson,rmarkdown,codetools,roxygen2,curl,rematch,R6,knitr,dplyr,bindr,automagic,commonmark,rprojroot,futile.options,semver,desc,stringi,Rcpp,harbor,remotes' from sysreqs online DB
#> INFO [2018-03-26 22:22:20] Adding CRAN packages: A3, abc, abc.data, assertthat, automagic, backports, bindr, bindrcpp, callr, cli, clisymbols, codetools, colorspace, commonmark, crayon, curl, data.table, debugme, desc, devtools, digest, dplyr, evaluate, formatR, fortunes, fs, futile.logger, futile.options, githubinstall, glue, highlight, htmltools, httr, jsonlite, knitr, lambda.r, lattice, locfit, magrittr, MASS, Matrix, MatrixModels, memoise, munsell, nnet, pacman, pbapply, pillar, pkgconfig, plyr, purrr, quantreg, R6, Rcpp, rematch, remotes, rjson, rlang, rmarkdown, roxygen2, rprojroot, rsconnect, rstudioapi, scales, semver, sessioninfo, SparseM, stringi, stringr, sys, tibble, whisker, withr, xml2, xtable, yaml
#> INFO [2018-03-26 22:22:20] Adding GitHub packages: r-hub/sysreqs@e4050e6068655ce519bb39f0508c7f10e19b6f0b, r-lib/pkgbuild@015f7f68ac4cae91ae7744558e4221034140f259, r-lib/pkgdown@5e4825875751c009444c56ce43d06324ec53e910, r-lib/pkgload@0cd9f8d9d455f447a0449513fa27745417289f73, r-lib/testthat@eda2d54a78e75e76f5544b78a14b50e88d926689
#> INFO [2018-03-26 22:22:20] Created Dockerfile-Object based on sessionInfo
#> INFO [2018-03-26 22:22:20] Writing dockerfile to /tmp/RtmpBJq2xz/file19e52b21e0b

print(readLines(temp))

#>  [1] "FROM rocker/r-ver:3.4.4"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#>  [2] "LABEL maintainer=\"daniel\""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#>  [3] "RUN export DEBIAN_FRONTEND=noninteractive; apt-get -y update \\"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
#>  [4] "  && apt-get install -y git-core \\"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    
#>  [5] "\tlibapparmor-dev \\"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   
#>  [6] "\tlibcurl4-openssl-dev \\"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              
#>  [7] "\tlibssl-dev \\"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
#>  [8] "\tlibxml2-dev \\"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       
#>  [9] "\tmake \\"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              
#> [10] "\tpandoc \\"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#> [11] "\tpandoc-citeproc"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#> [12] "RUN [\"install2.r\", \"A3\", \"abc\", \"abc.data\", \"assertthat\", \"automagic\", \"backports\", \"bindr\", \"bindrcpp\", \"callr\", \"cli\", \"clisymbols\", \"codetools\", \"colorspace\", \"commonmark\", \"crayon\", \"curl\", \"data.table\", \"debugme\", \"desc\", \"devtools\", \"digest\", \"dplyr\", \"evaluate\", \"formatR\", \"fortunes\", \"fs\", \"futile.logger\", \"futile.options\", \"githubinstall\", \"glue\", \"highlight\", \"htmltools\", \"httr\", \"jsonlite\", \"knitr\", \"lambda.r\", \"lattice\", \"locfit\", \"magrittr\", \"MASS\", \"Matrix\", \"MatrixModels\", \"memoise\", \"munsell\", \"nnet\", \"pacman\", \"pbapply\", \"pillar\", \"pkgconfig\", \"plyr\", \"purrr\", \"quantreg\", \"R6\", \"Rcpp\", \"rematch\", \"remotes\", \"rjson\", \"rlang\", \"rmarkdown\", \"roxygen2\", \"rprojroot\", \"rsconnect\", \"rstudioapi\", \"scales\", \"semver\", \"sessioninfo\", \"SparseM\", \"stringi\", \"stringr\", \"sys\", \"tibble\", \"whisker\", \"withr\", \"xml2\", \"xtable\", \"yaml\"]"
#> [13] "RUN [\"installGithub.r\", \"r-hub/sysreqs@e4050e6068655ce519bb39f0508c7f10e19b6f0b\", \"r-lib/pkgbuild@015f7f68ac4cae91ae7744558e4221034140f259\", \"r-lib/pkgdown@5e4825875751c009444c56ce43d06324ec53e910\", \"r-lib/pkgload@0cd9f8d9d455f447a0449513fa27745417289f73\", \"r-lib/testthat@eda2d54a78e75e76f5544b78a14b50e88d926689\"]"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> [14] "WORKDIR /payload/"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#> [15] "CMD [\"R\"]"