Write a Dockerfile object to a dockerfile

Write a Dockerfile object to a dockerfile

# S4 method for Dockerfile
write(x, file = file.path(getwd(), "Dockerfile"))

Arguments

x
Dockerfile object to be serialized

file
optional argument specifying a costum file path

Examples

# write a dockerfile with default parameters to temporary file and show content:
temp = tempfile()
write(dockerfile(), file=temp)

#> INFO [2019-08-20 16:44:09] Going online? TRUE  ... to retrieve system dependencies (sysreq-api)
#> INFO [2019-08-20 16:44:09] Trying to determine system requirements for the package(s) 'assertthat,automagic,backports,BiocManager,callr,cli,codetools,commonmark,crayon,curl,debugme,desc,devtools,digest,evaluate,fansi,formatR,fs,futile.logger,futile.options,glue,highlight,htmltools,httpuv,httr,jsonlite,knitr,lambda.r,later,magrittr,MASS,memoise,mime,miniUI,pillar,pkgbuild,pkgconfig,pkgdown,pkgload,prettyunits,processx,promises,ps,purrr,R6,Rcpp,rematch2,remotes,rjson,rlang,rmarkdown,roxygen2,rprojroot,rstudioapi,semver,sessioninfo,shiny,shinyFiles,stevedore,stringi,stringr,sysreqs,testthat,tibble,usethis,versions,whisker,withr,xfun,xml2,xtable,yaml' from sysreqs online DB
#> INFO [2019-08-20 16:44:13] Adding CRAN packages: assertthat, automagic, backports, BiocManager, callr, cli, codetools, commonmark, crayon, curl, debugme, desc, devtools, digest, evaluate, fansi, formatR, fs, futile.logger, futile.options, glue, highlight, htmltools, httpuv, httr, jsonlite, knitr, lambda.r, later, magrittr, MASS, memoise, mime, miniUI, pillar, pkgbuild, pkgconfig, pkgdown, pkgload, prettyunits, processx, promises, ps, purrr, R6, Rcpp, rematch2, remotes, rjson, rlang, rmarkdown, roxygen2, rprojroot, rstudioapi, semver, sessioninfo, shiny, shinyFiles, stevedore, stringi, stringr, testthat, tibble, usethis, versions, whisker, withr, xfun, xml2, xtable, yaml
#> INFO [2019-08-20 16:44:13] Adding GitHub packages: r-hub/sysreqs@3860f2b512a9c3bd3db6791c2ff467a1158f4048
#> INFO [2019-08-20 16:44:13] Created Dockerfile-Object based on sessionInfo

print(readLines(temp))

#>  [1] "FROM rocker/r-ver:3.6.1"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#>  [2] "LABEL maintainer=\"daniel\""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#>  [3] "RUN export DEBIAN_FRONTEND=noninteractive; apt-get -y update \\"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
#>  [4] "  && apt-get install -y git-core \\"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    
#>  [5] "\tlibcurl4-openssl-dev \\"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              
#>  [6] "\tlibssl-dev \\"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
#>  [7] "\tlibxml2-dev \\"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       
#>  [8] "\tmake \\"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              
#>  [9] "\tpandoc \\"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#> [10] "\tpandoc-citeproc"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#> [11] "RUN [\"install2.r\", \"assertthat\", \"automagic\", \"backports\", \"BiocManager\", \"callr\", \"cli\", \"codetools\", \"commonmark\", \"crayon\", \"curl\", \"debugme\", \"desc\", \"devtools\", \"digest\", \"evaluate\", \"fansi\", \"formatR\", \"fs\", \"futile.logger\", \"futile.options\", \"glue\", \"highlight\", \"htmltools\", \"httpuv\", \"httr\", \"jsonlite\", \"knitr\", \"lambda.r\", \"later\", \"magrittr\", \"MASS\", \"memoise\", \"mime\", \"miniUI\", \"pillar\", \"pkgbuild\", \"pkgconfig\", \"pkgdown\", \"pkgload\", \"prettyunits\", \"processx\", \"promises\", \"ps\", \"purrr\", \"R6\", \"Rcpp\", \"rematch2\", \"remotes\", \"rjson\", \"rlang\", \"rmarkdown\", \"roxygen2\", \"rprojroot\", \"rstudioapi\", \"semver\", \"sessioninfo\", \"shiny\", \"shinyFiles\", \"stevedore\", \"stringi\", \"stringr\", \"testthat\", \"tibble\", \"usethis\", \"versions\", \"whisker\", \"withr\", \"xfun\", \"xml2\", \"xtable\", \"yaml\"]"
#> [12] "RUN [\"installGithub.r\", \"r-hub/sysreqs@3860f2b512a9c3bd3db6791c2ff467a1158f4048\"]"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  
#> [13] "WORKDIR /payload/"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#> [14] "CMD [\"R\"]"